% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{scope_circle}
\alias{scope_circle}
\title{Border function}
\usage{
scope_circle(x = 0.3, y = 0.5, r = 0.3)
}
\arguments{
\item{x}{Origin of the circle}

\item{y}{Origin of the circle}

\item{r}{Radius of the circle}
}
\value{
A function that returns a vector of logicals
}
\description{
Border function
}
