% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{scope_grid}
\alias{scope_grid}
\title{Construct the data for a scopes plot}
\usage{
scope_grid(
  seed = 272,
  grain = 2000,
  border = scope_circle(),
  inner = scope_texture(frequency = 5),
  outer = scope_texture(frequency = 2),
  palette = viridis::viridis
)
}
\arguments{
\item{seed}{Random number generator seed}

\item{grain}{Number of pixels}

\item{border}{Function specifying the border}

\item{inner}{List of texture parameters for inner region}

\item{outer}{List of texture parameters for outer region}

\item{palette}{Palette generating function}
}
\value{
A tibble
}
\description{
Construct the data for a scopes plot
}
