% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders.R
\name{scope_hex}
\alias{scope_hex}
\title{Border function}
\usage{
scope_hex(x = 0.3, y = 0.5, r = 0.3)
}
\arguments{
\item{x}{Origin of the hexagon}

\item{y}{Origin of the hexagon}

\item{r}{(Long) radius of the hexagon}
}
\value{
A function that returns a vector of logicals
}
\description{
Border function
}
