% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{scope_plot}
\alias{scope_plot}
\title{Write scopes data to a png}
\usage{
scope_plot(scope, filename = NULL, grain = 2000)
}
\arguments{
\item{scope}{The scope data}

\item{filename}{Path to file (if default, use current graphics device)}

\item{grain}{Image width & height}
}
\description{
Write scopes data to a png
}
