% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{scope_texture}
\alias{scope_texture}
\title{Parameters for the fractal texture}
\usage{
scope_texture(
  noise = ambient::gen_worley,
  frequency = 1,
  fractal = ambient::billow,
  octaves = 3,
  ...
)
}
\arguments{
\item{noise}{Function generating the multidimensional noise}

\item{frequency}{Frequency}

\item{fractal}{Fractal function}

\item{octaves}{How many octaves}

\item{...}{Other arguments}
}
\value{
A list
}
\description{
Parameters for the fractal texture
}
