% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{scope_circle}
\alias{scope_circle}
\title{Border function}
\usage{
scope_circle(scope, origin_x = 0.3, origin_y = 0.5, radius = 0.3)
}
\arguments{
\item{scope}{The scope object}

\item{origin_x}{Origin of the circle}

\item{origin_y}{Origin of the circle}

\item{radius}{Radius of the circle}
}
\value{
vector of logicals
}
\description{
Border function
}
